# tixmoapi2/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "TixMo API",
    "version": "0.1.0",
    "description": "Enterprise-grade ticket vendor platform API for event promoters",
    "contact": {
      "name": "TixMo Team",
      "email": "api@tixmo.com"
    },
    "license": {
      "name": "Proprietary"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api/v1",
      "description": "Development server"
    },
    {
      "url": "https://api.tixmo.com/v1",
      "description": "Production server (future)"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter your JWT access token"
      }
    },
...
  },
...
  "paths": {
...
    "/analytics/events": {
      "get": {
        "summary": "Get event analytics",
        "tags": [
          "Analytics"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "startDate",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "in": "query",
            "name": "endDate",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "in": "query",
            "name": "organizationId",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Event analytics retrieved successfully"
          }
        }
      }
    },
...
    "/events": {
      "post": {
        "summary": "Create a new event",
        "description": "Create a new event for an organization. Requires ADMIN or PROMOTER role.",
        "tags": [
          "Events"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "title",
                  "description",
                  "organizationId",
                  "venueId",
                  "startDateTime",
                  "endDateTime"
                ],
                "properties": {
                  "title": {
                    "type": "string",
                    "example": "Summer Music Festival 2025"
                  },
                  "description": {
                    "type": "string",
                    "example": "Join us for an amazing night of live music"
                  },
                  "organizationId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "venueId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "startDateTime": {
                    "type": "string",
                    "format": "date-time",
                    "example": "2025-07-15T19:00:00Z"
                  },
                  "endDateTime": {
                    "type": "string",
                    "format": "date-time",
                    "example": "2025-07-15T23:00:00Z"
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "DRAFT",
                      "PUBLISHED",
                      "CANCELLED"
                    ],
                    "default": "DRAFT"
                  },
                  "capacity": {
                    "type": "integer",
                    "example": 5000
                  },
                  "imageUrl": {
                    "type": "string",
                    "format": "uri"
                  },
                  "metadata": {
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Event created successfully"
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Organization or venue not found"
          }
        }
      },
      "get": {
        "summary": "List all events",
        "description": "Get a paginated list of events with optional filters",
        "tags": [
          "Events"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 20
            }
          },
          {
            "in": "query",
            "name": "organizationId",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "in": "query",
            "name": "venueId",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string",
              "enum": [
                "DRAFT",
                "PUBLISHED",
                "CANCELLED"
              ]
            }
          },
          {
            "in": "query",
            "name": "search",
            "schema": {
              "type": "string"
            },
            "description": "Search in title and description"
          }
        ],
        "responses": {
          "200": {
            "description": "Events retrieved successfully"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/events/{id}": {
      "get": {
        "summary": "Get event by ID",
        "description": "Get detailed information about an event including venue and ticket counts",
        "tags": [
          "Events"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Event retrieved successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Event not found"
          }
        }
      },
      "put": {
        "summary": "Update event",
        "description": "Update event details",
        "tags": [
          "Events"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "venueId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "startDateTime": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "endDateTime": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "DRAFT",
                      "PUBLISHED",
                      "CANCELLED"
                    ]
                  },
                  "capacity": {
                    "type": "integer"
                  },
                  "imageUrl": {
                    "type": "string"
                  },
                  "metadata": {
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Event updated successfully"
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Event not found"
          }
        }
      },
      "delete": {
        "summary": "Delete event",
        "description": "Delete an event (admin only). Cannot delete if event has tickets.",
        "tags": [
          "Events"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Event deleted successfully"
          },
          "400": {
            "description": "Cannot delete event with existing tickets"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - Admin only"
          },
          "404": {
            "description": "Event not found"
          }
        }
      }
    },
    "/events/{id}/publish": {
      "post": {
        "summary": "Publish event",
        "description": "Change event status from DRAFT to PUBLISHED",
        "tags": [
          "Events"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Event published successfully"
          },
          "400": {
            "description": "Event is already published or cancelled"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Event not found"
          }
        }
      }
    },
    "/events/{id}/cancel": {
      "post": {
        "summary": "Cancel event",
        "description": "Change event status to CANCELLED",
        "tags": [
          "Events"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Event cancelled successfully"
          },
          "400": {
            "description": "Event is already cancelled"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Event not found"
          }
        }
      }
    },
    "/events/{id}/stats": {
      "get": {
        "summary": "Get event statistics",
        "description": "Get real-time statistics for an event including tickets, scans, and occupancy",
        "tags": [
          "Events"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Event statistics retrieved successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Event not found"
          }
        }
      }
    },
    "/events/{id}/occupancy": {
      "get": {
        "summary": "Get current occupancy",
        "description": "Get real-time venue occupancy for the event",
        "tags": [
          "Events"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Occupancy data retrieved successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Event not found"
          }
        }
      }
    },
    "/events/{id}/timeline": {
      "get": {
        "summary": "Get entry timeline",
        "description": "Get hourly breakdown of entries and exits",
        "tags": [
          "Events"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "in": "query",
            "name": "hours",
            "schema": {
              "type": "integer",
              "default": 24
            },
            "description": "Number of hours to look back"
          }
        ],
        "responses": {
          "200": {
            "description": "Entry timeline retrieved successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Event not found"
          }
        }
      }
    },
    "/events/{id}/scanner-stats": {
      "get": {
        "summary": "Get scanner statistics",
        "description": "Get statistics about scanner activity for the event",
        "tags": [
          "Events"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Scanner statistics retrieved successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Event not found"
          }
        }
      }
    },
...
  }
}